Switched back to generating all vertices Also the world gets drawn from near to far(using a snake algorithm)

+++ include/craftus/world/blocks.h
@@ -3,8 +3,10 @@
 
 #include "craftus/world/direction.h"
 
+#include <stdint.h>
+
 typedef enum { Block_Air = 0, Block_Stone, Block_Dirt, Block_Grass } Block;
 
-void Blocks_GetUVs(Block block, Direction side, float* out);
+void Blocks_GetUVs(Block block, Direction side, uint8_t* out);
 
-#endif  // !BLOCK_H_INCLUDED 
+#endif  // !BLOCK_H_INCLUDED
